All: Resolved a couple of problems with language (enUS, deDE, ...)  * forcedLanguage was overwritten by client data.  * Partially code rewrite to prevent unexpected behavior for LogonCommServer/Client (server/logon conversations)
All: Resolved problems with NpcScriptText.  If packet SMSG_MESSAGECHAT gets build, we check if target Object (Player) has a forcedLanguage set to its account. If there is a forcedLanguage we check for localized creature name and localized script text and use it in our packet.

This commit will partially resolve issue #110

Sorry for this "large" commit.