Style/Rewrite Instance class
* use class member initialization
* move db load/delete/update stuff to InstanceMgr class
* split code to seperate file

Speaking of changes, the design was pretty weird. The legacy code changes the database without updating the saved data. Over all we have the class InstanceMgr which should take care about the data management.
There is still a lot to do in InstanceMgr since it is not thread save, we should use modern singleton implementation to get rid of some logic problems.