Introduce multiversion OpcodeTables
Since multiversion was introduced here we end up having a lot of version specific code and definitions. So called Opcodes are defined with hex values for every specific version.
The new solution replace these version specific values with an internal index and gives back the hex value from a std::map.

It is currently a draft! Do not use it ;-)
I try to expand the functionality to cover most user stories as possible.