Remove old DB check function
V530 The return value of function 'getItemDisplayIdForEntry' is required to be utilized. world MySQLDataStore.cpp 4082
V530 The return value of function 'getItemDisplayIdForEntry' is required to be utilized. world MySQLDataStore.cpp 4084
V530 The return value of function 'getItemDisplayIdForEntry' is required to be utilized. world MySQLDataStore.cpp 4086
V530 The return value of function 'getItemDisplayIdForEntry' is required to be utilized. world MySQLDataStore.cpp 4103
V530 The return value of function 'getItemDisplayIdForEntry' is required to be utilized. world MySQLDataStore.cpp 4105
V530 The return value of function 'getItemDisplayIdForEntry' is required to be utilized. world MySQLDataStore.cpp 4107