Clear up more asserts
NOTE: there are a few valid cases to assert the program e.g. missing config values. Let's move on and do not implement ASSERTS to check out unexpected behavior. If you need to make sure that some criteria has to be reached, just use if/else statements (works in most cases). Just ask the following question "is it necessary to stop the program or is it enough to  further develop the issue".
Pretty sure I broke some thing e.g. Transports.